requirements.txt
aiogram==2.7
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
Babel==2.14.0
certifi==2023.11.17
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
python-dotenv==1.0.0
yarl==1.9.4
